Moved KVS to separate module with separate test harness. Will test this in isolation first before integrating with Registry
